refactor: Rename page (#64)
